Removed old parsing stuff and refactored tests to match new way
